rename Client to WalletApiWrapper, WalletClient to WalletApi, TestClient to Client
