+++ src/main/java/org/tron/common/crypto/Hash.java
@@ -27,7 +27,7 @@ import java.security.Security;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.tron.common.crypto.jce.TronCastleProvider;
-import org.tron.walletserver.WalletClient;
+import org.tron.walletserver.WalletApi;
 import org.tron.common.utils.ByteArray;
 
 public class Hash {
@@ -129,7 +129,7 @@ public class Hash {
   public static byte[] sha3omit12(byte[] input) {
     byte[] hash = sha3(input);
     byte[] address = copyOfRange(hash, 11, hash.length);
-    address[0] = WalletClient.getAddressPreFixByte();
+    address[0] = WalletApi.getAddressPreFixByte();
     return address;
   }
 }
