+++ src/main/java/org/tron/common/utils/Utils.java
@@ -84,7 +84,7 @@ import org.tron.protos.Protocol.TransactionInfo;
 import org.tron.protos.Protocol.TransactionInfo.Log;
 import org.tron.protos.Protocol.Vote;
 import org.tron.protos.Protocol.Witness;
-import org.tron.walletserver.WalletClient;
+import org.tron.walletserver.WalletApi;
 
 public class Utils {
 
@@ -127,7 +127,7 @@ public class Utils {
   public static String printAccount(Account account) {
     String result = "";
     result += "address: ";
-    result += WalletClient.encode58Check(account.getAddress().toByteArray());
+    result += WalletApi.encode58Check(account.getAddress().toByteArray());
     result += "\n";
     if (account.getAccountId() != null && !account.getAccountId().isEmpty()) {
       result += "account_id: ";
@@ -179,7 +179,7 @@ public class Utils {
         result += "{";
         result += "\n";
         result += "  vote_address: ";
-        result += WalletClient.encode58Check(vote.getVoteAddress().toByteArray());
+        result += WalletApi.encode58Check(vote.getVoteAddress().toByteArray());
         result += "\n";
         result += "  vote_count: ";
         result += vote.getVoteCount();
@@ -318,7 +318,7 @@ public class Utils {
   public static String printWitness(Witness witness) {
     String result = "";
     result += "address: ";
-    result += WalletClient.encode58Check(witness.getAddress().toByteArray());
+    result += WalletApi.encode58Check(witness.getAddress().toByteArray());
     result += "\n";
     result += "voteCount: ";
     result += witness.getVoteCount();
@@ -366,7 +366,7 @@ public class Utils {
     result += "\n";
     result += "approvalsList: [ \n";
     for (ByteString address : proposal.getApprovalsList()) {
-      result += WalletClient.encode58Check(address.toByteArray());
+      result += WalletApi.encode58Check(address.toByteArray());
       result += "\n";
     }
     result += "]";
@@ -421,7 +421,7 @@ public class Utils {
   public static String printAssetIssue(AssetIssueContract assetIssue) {
     String result = "";
     result += "owner_address: ";
-    result += WalletClient.encode58Check(assetIssue.getOwnerAddress().toByteArray());
+    result += WalletApi.encode58Check(assetIssue.getOwnerAddress().toByteArray());
     result += "\n";
     result += "name: ";
     result += new String(assetIssue.getName().toByteArray(), Charset.forName("UTF-8"));
@@ -520,12 +520,12 @@ public class Utils {
           if (accountCreateContract.getAccountAddress() != null
               && !accountCreateContract.getAccountAddress().isEmpty()) {
             result += "account_address: ";
-            result += WalletClient
+            result += WalletApi
                 .encode58Check(accountCreateContract.getAccountAddress().toByteArray());
             result += "\n";
           }
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(accountCreateContract.getOwnerAddress().toByteArray());
           result += "\n";
           break;
@@ -540,7 +540,7 @@ public class Utils {
             result += "\n";
           }
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(accountUpdateContract.getOwnerAddress().toByteArray());
           result += "\n";
           break;
@@ -548,11 +548,11 @@ public class Utils {
           TransferContract transferContract = contract.getParameter()
               .unpack(TransferContract.class);
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(transferContract.getOwnerAddress().toByteArray());
           result += "\n";
           result += "to_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(transferContract.getToAddress().toByteArray());
           result += "\n";
           result += "amount: ";
@@ -567,11 +567,11 @@ public class Utils {
               Charset.forName("UTF-8"));
           result += "\n";
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(transferAssetContract.getOwnerAddress().toByteArray());
           result += "\n";
           result += "to_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(transferAssetContract.getToAddress().toByteArray());
           result += "\n";
           result += "amount: ";
@@ -586,7 +586,7 @@ public class Utils {
           VoteWitnessContract voteWitnessContract = contract.getParameter()
               .unpack(VoteWitnessContract.class);
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(voteWitnessContract.getOwnerAddress().toByteArray());
           result += "\n";
           result += "votes: ";
@@ -597,7 +597,7 @@ public class Utils {
             result += "[";
             result += "\n";
             result += "vote_address: ";
-            result += WalletClient
+            result += WalletApi
                 .encode58Check(vote.getVoteAddress().toByteArray());
             result += "\n";
             result += "vote_count: ";
@@ -613,7 +613,7 @@ public class Utils {
           WitnessCreateContract witnessCreateContract = contract.getParameter()
               .unpack(WitnessCreateContract.class);
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(witnessCreateContract.getOwnerAddress().toByteArray());
           result += "\n";
           result += "url: ";
@@ -625,7 +625,7 @@ public class Utils {
           WitnessUpdateContract witnessUpdateContract = contract.getParameter()
               .unpack(WitnessUpdateContract.class);
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(witnessUpdateContract.getOwnerAddress().toByteArray());
           result += "\n";
           result += "url: ";
@@ -642,7 +642,7 @@ public class Utils {
           UpdateAssetContract updateAssetContract = contract.getParameter()
               .unpack(UpdateAssetContract.class);
           result += "owner_address: ";
-          result += WalletClient.encode58Check(updateAssetContract.getOwnerAddress().toByteArray());
+          result += WalletApi.encode58Check(updateAssetContract.getOwnerAddress().toByteArray());
           result += "\n";
           result += "description: ";
           result += new String(updateAssetContract.getDescription().toByteArray(),
@@ -667,11 +667,11 @@ public class Utils {
               Charset.forName("UTF-8"));
           result += "\n";
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(participateAssetIssueContract.getOwnerAddress().toByteArray());
           result += "\n";
           result += "to_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(participateAssetIssueContract.getToAddress().toByteArray());
           result += "\n";
           result += "amount: ";
@@ -682,7 +682,7 @@ public class Utils {
           FreezeBalanceContract freezeBalanceContract = contract.getParameter()
               .unpack(FreezeBalanceContract.class);
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(freezeBalanceContract.getOwnerAddress().toByteArray());
           result += "\n";
           result += "frozen_balance: ";
@@ -696,7 +696,7 @@ public class Utils {
           UnfreezeBalanceContract unfreezeBalanceContract = contract.getParameter()
               .unpack(UnfreezeBalanceContract.class);
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(unfreezeBalanceContract.getOwnerAddress().toByteArray());
           result += "\n";
           break;
@@ -704,7 +704,7 @@ public class Utils {
           UnfreezeAssetContract unfreezeAssetContract = contract.getParameter()
               .unpack(UnfreezeAssetContract.class);
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(unfreezeAssetContract.getOwnerAddress().toByteArray());
           result += "\n";
           break;
@@ -712,7 +712,7 @@ public class Utils {
           WithdrawBalanceContract withdrawBalanceContract = contract.getParameter()
               .unpack(WithdrawBalanceContract.class);
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(withdrawBalanceContract.getOwnerAddress().toByteArray());
           result += "\n";
           break;
@@ -721,7 +721,7 @@ public class Utils {
               .unpack(CreateSmartContract.class);
           SmartContract newContract = createSmartContract.getNewContract();
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(createSmartContract.getOwnerAddress().toByteArray());
           result += "\n";
           result += "ABI: ";
@@ -734,18 +734,18 @@ public class Utils {
           result += newContract.getCallValue();
           result += "\n";
           result += "contract_address:";
-          result += WalletClient.encode58Check(newContract.getContractAddress().toByteArray());
+          result += WalletApi.encode58Check(newContract.getContractAddress().toByteArray());
           result += "\n";
           break;
         case TriggerSmartContract:
           TriggerSmartContract triggerSmartContract = contract.getParameter()
               .unpack(TriggerSmartContract.class);
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(triggerSmartContract.getOwnerAddress().toByteArray());
           result += "\n";
           result += "contract_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(triggerSmartContract.getContractAddress().toByteArray());
           result += "\n";
           result += "call_value:";
@@ -759,7 +759,7 @@ public class Utils {
           ProposalCreateContract proposalCreateContract = contract.getParameter()
               .unpack(ProposalCreateContract.class);
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(proposalCreateContract.getOwnerAddress().toByteArray());
           result += "\n";
           result += "parametersMap: ";
@@ -770,7 +770,7 @@ public class Utils {
           ProposalApproveContract proposalApproveContract = contract.getParameter()
               .unpack(ProposalApproveContract.class);
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(proposalApproveContract.getOwnerAddress().toByteArray());
           result += "\n";
           result += "proposal id: ";
@@ -784,14 +784,14 @@ public class Utils {
           ProposalDeleteContract proposalDeleteContract = contract.getParameter()
               .unpack(ProposalDeleteContract.class);
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(proposalDeleteContract.getOwnerAddress().toByteArray());
           break;
         case BuyStorageContract:
           BuyStorageContract buyStorageContract = contract.getParameter()
               .unpack(BuyStorageContract.class);
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(buyStorageContract.getOwnerAddress().toByteArray());
           result += "\n";
           result += "quant:";
@@ -802,7 +802,7 @@ public class Utils {
           SellStorageContract sellStorageContract = contract.getParameter()
               .unpack(SellStorageContract.class);
           result += "owner_address: ";
-          result += WalletClient
+          result += WalletApi
               .encode58Check(sellStorageContract.getOwnerAddress().toByteArray());
           result += "\n";
           result += "storageBytes:";
@@ -1018,7 +1018,7 @@ public class Utils {
     result += "\n";
     result += "contractAddress: ";
     result += "\n";
-    result += WalletClient.encode58Check(transactionInfo.getContractAddress().toByteArray());
+    result += WalletApi.encode58Check(transactionInfo.getContractAddress().toByteArray());
     result += "\n";
     result += "logList: ";
     result += "\n";
@@ -1157,7 +1157,7 @@ public class Utils {
     result += "\n";
 
     result += "witness_address: ";
-    result += WalletClient.encode58Check(raw.getWitnessAddress().toByteArray());
+    result += WalletApi.encode58Check(raw.getWitnessAddress().toByteArray());
     result += "\n";
 
     return result;
@@ -1389,7 +1389,7 @@ public class Utils {
         StringUtils.clear(passwd0);
         StringUtils.clear(passwd1);
       }
-      if (WalletClient.passwordValid(password)) {
+      if (WalletApi.passwordValid(password)) {
         return password;
       }
       if (!checkStrength) {
