+++ src/main/java/org/tron/keystore/Wallet.java
@@ -1,6 +1,5 @@
 package org.tron.keystore;
 
-import static java.nio.charset.StandardCharsets.UTF_8;
 import java.security.InvalidAlgorithmParameterException;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
@@ -21,7 +20,7 @@ import org.tron.common.crypto.ECKey;
 import org.tron.common.crypto.Hash;
 import org.tron.common.utils.ByteArray;
 import org.tron.core.exception.CipherException;
-import org.tron.walletserver.WalletClient;
+import org.tron.walletserver.WalletApi;
 
 /**
  * <p>Ethereum wallet file management. For reference, refer to <a href="https://github.com/ethereum/wiki/wiki/Web3-Secret-Storage-Definition">
@@ -94,7 +93,7 @@ public class Wallet {
       int n, int p) {
 
     WalletFile walletFile = new WalletFile();
-    walletFile.setAddress(WalletClient.encode58Check(ecKeyPair.getAddress()));
+    walletFile.setAddress(WalletApi.encode58Check(ecKeyPair.getAddress()));
 
     WalletFile.Crypto crypto = new WalletFile.Crypto();
     crypto.setCipher(CIPHER);
